Issue #16729: Improve SuppressionSingleFilter examples to demonstrate columns and id properties
