+++ src/xdocs-examples/java/com/puppycrawl/tools/checkstyle/filters/SuppressionSingleFilterExamplesTest.java
@@ -19,23 +19,15 @@
 
 package com.puppycrawl.tools.checkstyle.filters;
 
-import static com.puppycrawl.tools.checkstyle.checks.coding.EqualsAvoidNullCheck.MSG_EQUALS_AVOID_NULL;
 import static com.puppycrawl.tools.checkstyle.checks.coding.MagicNumberCheck.MSG_KEY;
-import static com.puppycrawl.tools.checkstyle.checks.javadoc.JavadocMethodCheck.MSG_RETURN_EXPECTED;
 import static com.puppycrawl.tools.checkstyle.checks.regexp.RegexpCheck.MSG_ILLEGAL_REGEXP;
 
 import org.junit.jupiter.api.Test;
 
 import com.puppycrawl.tools.checkstyle.AbstractExamplesModuleTestSupport;
-import com.puppycrawl.tools.checkstyle.checks.coding.EqualsAvoidNullCheck;
 import com.puppycrawl.tools.checkstyle.checks.coding.MagicNumberCheck;
-import com.puppycrawl.tools.checkstyle.checks.javadoc.JavadocMethodCheck;
-import com.puppycrawl.tools.checkstyle.checks.naming.ConstantNameCheck;
 import com.puppycrawl.tools.checkstyle.checks.naming.MemberNameCheck;
-import com.puppycrawl.tools.checkstyle.checks.naming.MethodNameCheck;
 import com.puppycrawl.tools.checkstyle.checks.regexp.RegexpCheck;
-import com.puppycrawl.tools.checkstyle.checks.sizes.FileLengthCheck;
-import com.puppycrawl.tools.checkstyle.checks.sizes.ParameterNumberCheck;
 import com.puppycrawl.tools.checkstyle.checks.whitespace.NoWhitespaceAfterCheck;
 
 public class SuppressionSingleFilterExamplesTest extends AbstractExamplesModuleTestSupport {
@@ -58,10 +50,15 @@ public class SuppressionSingleFilterExamplesTest extends AbstractExamplesModuleT
     @Test
     public void testExample2() throws Exception {
         final String[] expectedWithoutFilter = {
-            "19:13: " + getCheckMessage(EqualsAvoidNullCheck.class, MSG_EQUALS_AVOID_NULL),
-            "25: " + getCheckMessage(JavadocMethodCheck.class, MSG_RETURN_EXPECTED),
+            "19:12: " + getCheckMessage(NoWhitespaceAfterCheck.class,
+                    NoWhitespaceAfterCheck.MSG_KEY, "."),
+            "23:9: " + getCheckMessage(NoWhitespaceAfterCheck.class,
+                    NoWhitespaceAfterCheck.MSG_KEY, "int"),
         };
-        final String[] expectedWithFilter = {};
+        final String[] expectedWithFilter = {
+            "23:9: " + getCheckMessage(NoWhitespaceAfterCheck.class,
+                    NoWhitespaceAfterCheck.MSG_KEY, "int"),
+        };
 
         verifyFilterWithInlineConfigParser(getPath("Example2.java"),
                 expectedWithoutFilter, expectedWithFilter);
@@ -82,90 +79,12 @@ public class SuppressionSingleFilterExamplesTest extends AbstractExamplesModuleT
     @Test
     public void testExample4() throws Exception {
         final String[] expectedWithoutFilter = {
-            "18:12: " + getCheckMessage(NoWhitespaceAfterCheck.class,
-                    NoWhitespaceAfterCheck.MSG_KEY, "."),
-            "22:9: " + getCheckMessage(NoWhitespaceAfterCheck.class,
-                    NoWhitespaceAfterCheck.MSG_KEY, "int"),
+            "18:15: " + getCheckMessage(MemberNameCheck.class, "name.invalidPattern",
+                    "MyVariable", "^[a-z][a-zA-Z0-9]*$"),
         };
         final String[] expectedWithFilter = {};
 
         verifyFilterWithInlineConfigParser(getPath("Example4.java"),
-                expectedWithoutFilter, expectedWithFilter);
-    }
-
-    @Test
-    public void testExample5() throws Exception {
-        final String[] expectedWithoutFilter = {
-            "14:15: " + getCheckMessage(MethodNameCheck.class, "name.invalidPattern",
-                    "example_Method", "^[a-z][a-zA-Z0-9]*$"),
-            "17:15: " + getCheckMessage(MethodNameCheck.class, "name.invalidPattern",
-                    "Another_Method", "^[a-z][a-zA-Z0-9]*$"),
-        };
-        final String[] expectedWithFilter = {};
-
-        verifyFilterWithInlineConfigParser(getPath("Example5.java"),
-                expectedWithoutFilter, expectedWithFilter);
-    }
-
-    @Test
-    public void testExample6() throws Exception {
-        final String[] expectedWithoutFilter = {
-            "14:28: " + getCheckMessage(ConstantNameCheck.class, "name.invalidPattern",
-                    "myConstant", "^[A-Z][A-Z0-9]*(_[A-Z0-9]+)*$"),
-        };
-        final String[] expectedWithFilter = {};
-
-        verifyFilterWithInlineConfigParser(getPath("Example6.java"),
-                expectedWithoutFilter, expectedWithFilter);
-    }
-
-    @Test
-    public void testExample7() throws Exception {
-        final String[] expectedWithoutFilter = {
-            "15:15: " + getCheckMessage(MemberNameCheck.class, "name.invalidPattern",
-                    "MyVariable", "^[a-z][a-zA-Z0-9]*$"),
-            "17:15: " + getCheckMessage(MethodNameCheck.class, "name.invalidPattern",
-                    "MyMethod", "^[a-z][a-zA-Z0-9]*$"),
-        };
-        final String[] expectedWithFilter = {};
-
-        verifyFilterWithInlineConfigParser(getPath("Example7.java"),
-                expectedWithoutFilter, expectedWithFilter);
-    }
-
-    @Test
-    public void testExample8() throws Exception {
-        final String[] expectedWithoutFilter = {
-            "18:15: " + getCheckMessage(ParameterNumberCheck.class,
-                    ParameterNumberCheck.MSG_KEY, 5, 6),
-        };
-        final String[] expectedWithFilter = {};
-
-        verifyFilterWithInlineConfigParser(getPath("Example8.java"),
-                expectedWithoutFilter, expectedWithFilter);
-    }
-
-    @Test
-    public void testExample9() throws Exception {
-        final String[] expectedWithoutFilter = {
-            "1: " + getCheckMessage(FileLengthCheck.class,
-                    FileLengthCheck.MSG_KEY, 21, 1),
-        };
-        final String[] expectedWithFilter = {};
-
-        verifyFilterWithInlineConfigParser(getPath("Example9.java"),
-                expectedWithoutFilter, expectedWithFilter);
-    }
-
-    @Test
-    public void testExample10() throws Exception {
-        final String[] expectedWithoutFilter = {
-            "16:18: " + getCheckMessage(MemberNameCheck.class, "name.invalidPattern",
-                    "log", "^[A-Z][a-zA-Z0-9]*$"),
-        };
-        final String[] expectedWithFilter = {};
-
-        verifyFilterWithInlineConfigParser(getPath("Example10.java"),
                 expectedWithoutFilter, expectedWithFilter);
     }
 }
