--- src/xdocs-examples/resources/com/puppycrawl/tools/checkstyle/filters/suppressionsinglefilter/Example9.java
@@ -1,21 +1,0 @@
-/*xml
-<module name="Checker">
-  <module name="TreeWalker">
-    <module name="FileLength">
-      <property name="max" value="1"/>
-    </module>
-  </module>
-  <module name="SuppressionSingleFilter">
-    <property name="files" value="Example9.java"/>
-    <property name="checks" value="FileLength"/>
-  </module>
-</module>
-*/
-package com.puppycrawl.tools.checkstyle.filters.suppressionsinglefilter;
-// filtered violation 14 lines above 'File length is 21 lines (max allowed is 1)'
-// xdoc section -- start
-/* filtered violation on 1st line  'File length is 4 lines (max allowed is 1)' */
-public class Example9 {
-
-}
-// xdoc section -- end
